update layout for seach profile
