+++ src/app/(authorized_layout)/find/page.tsx
@@ -1,0 +1,11 @@
+import SearchResults from 'src/modules/search/components/SearchResults';
+import Search from 'src/modules/search/components/Search';
+
+export default async function Home() {
+  return (
+    <>
+      <Search />
+      <SearchResults />
+    </>
+  );
+}
