--- src/app/(authorized_layout)/search/page.tsx
@@ -1,5 +1,0 @@
-import SeachPage from 'src/modules/search/components/SearchPage';
-
-export default function Home() {
-  return <SeachPage />;
-}
