+++ src/modules/share/components/_dashboard/user/profile/WorkForm.tsx
@@ -69,7 +69,7 @@ const WorkForm = ({ defaultValues, onSave }: any) => {
           type: 'text',
           placeholder: 'yyyy-yyyy',
         },
-      ].map((item) => {
+      ].map(item => {
         return item.name !== 'startDate' && item.name !== 'endDate' ? (
           <Controller
             key={item.name}
@@ -111,7 +111,7 @@ const WorkForm = ({ defaultValues, onSave }: any) => {
                   onChange={(event: any) => {
                     onChange(event);
                   }}
-                  renderInput={(params) => (
+                  renderInput={params => (
                     <TextField
                       margin="normal"
                       {...register(item.name, { required: false })}
